@@ -16,44 +16,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/cdk": "^7.1.1",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.1.10",
-    "adm-zip": "^0.4.11",
-    "bootstrap": "^4.1.3",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^7.2.2",
+    "@angular/cdk": "^7.2.2",
+    "@angular/common": "^7.2.2",
+    "@angular/compiler": "^7.2.2",
+    "@angular/core": "^7.2.2",
+    "@angular/forms": "^7.2.2",
+    "@angular/http": "^7.2.2",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.2",
+    "@angular/platform-browser-dynamic": "^7.2.2",
+    "@angular/router": "^7.2.2",
+    "adm-zip": "^0.4.13",
+    "bootstrap": "^4.3.1",
+    "core-js": "^2.6.10",
     "rxjs": "^6.3.3",
-    "rxjs-compat": "^6.3.3",
-    "webpack-dev-server": "^3.1.11",
-    "zone.js": "^0.8.26"
+    "rxjs-compat": "^6.5.3",
+    "webpack-dev-server": "^3.8.2",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^6.0.3",
     "@angular-devkit/build-angular": "~0.6.8",
-    "typescript": "~2.7.2",
-    "@angular/cli": "~6.0.8",
-    "@angular/language-service": "^6.0.3",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular/cli": "^7.3.9",
+    "@angular/compiler": "^7.2.2",
+    "@angular/compiler-cli": "^7.2.15",
+    "@angular/language-service": "7.2.2",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.8",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.2.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
-    "tslint": "~5.9.1"
+    "tslint": "~5.9.1",
+    "typescript": "^3.6.4"
   }
 }
